migrate layout to compose
