+++ app/build.gradle
@@ -12,7 +12,7 @@ android {
     defaultConfig {
         applicationId "com.example.mymovieslist"
         minSdk 21
-        targetSdk 31
+        targetSdk 32
         versionCode 1
         versionName "1.0"
 
@@ -80,7 +80,11 @@ dependencies {
     testImplementation 'junit:junit:4.13.2'
     androidTestImplementation 'androidx.test.ext:junit:4.13.2'
 
-    implementation "com.github.skydoves:landscapist-glide:1.5.3"
+
+    implementation "androidx.compose.material:material-icons-extended:1.3.0-alpha01"
+    implementation "androidx.compose.material3:material3-window-size-class:1.0.0-alpha14"
+
+    implementation "com.github.skydoves:landscapist-coil:1.5.0"
     implementation "com.google.accompanist:accompanist-swiperefresh:0.24.13-rc"
     implementation "com.google.accompanist:accompanist-systemuicontroller:0.24.13-rc"
     implementation "com.google.android.material:compose-theme-adapter:1.1.1"
@@ -100,4 +104,16 @@ dependencies {
 
 kapt {
     correctErrorTypes true
+}
+
+tasks.withType(org.jetbrains.kotlin.gradle.tasks.KotlinCompile).configureEach {
+    kotlinOptions {
+        // Treat all Kotlin warnings as errors (disabled by default)
+        allWarningsAsErrors = project.hasProperty("warningsAsErrors") ? project.warningsAsErrors : false
+
+        freeCompilerArgs += '-opt-in=kotlin.RequiresOptIn'
+
+        // Set JVM target to 1.8
+        jvmTarget = "1.8"
+    }
 }
